Создал в папке Servises файл QuestionFactory и вынес туда весь массив с данными и логику передачи данных для вопроса в MovieQuizViewController реализовал логку передачи таким образом что бы вопросы не повторялись, все структуры перенес в папку Models в соответствующие файлы, в MovieQuizViewController оптимизировал код создав отдельную функцию для создания вопроса
