+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,18 @@
+//
+//  QuizQuestion.swift
+//  MovieQuiz
+//
+//  Created by Константин Филиппов on 20.06.2024.
+//
+
+import Foundation
+
+// структура вопроса
+
+struct QuizQuestion {
+    
+    let image: String
+    let text: String
+    let correctAnswer: Bool
+    
+}
