+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,18 @@
+//
+//  QuizStepViewModel.swift
+//  MovieQuiz
+//
+//  Created by Константин Филиппов on 20.06.2024.
+//
+
+import UIKit
+
+// для состояния "Вопрос показан"
+
+struct QuizStepViewModel {
+    
+    let image: UIImage
+    let question: String
+    let questionNumber: String
+    
+}
